[FIX] CSRF

+++ forum/admin_ranks.php
@@ -3,7 +3,7 @@
  * @copyright   &copy; 2005-2026 PHPBoost
  * @license     https://www.gnu.org/licenses/gpl-3.0.html GNU/GPL-3.0
  * @author      Regis VIARRE <crowkait@phpboost.com>
- * @version     PHPBoost 6.0 - last update: 2021 12 01
+ * @version     PHPBoost 6.0 - last update: 2026 08 22
  * @since       PHPBoost 1.2 - 2005 10 30
  * @contributor Julien BRISWALTER <j1.seth@phpboost.com>
  * @contributor Arnaud GENET <elenwii@phpboost.com>
@@ -18,6 +18,7 @@ define('TITLE', $lang['forum.ranks.management']);
 require_once('../admin/admin_header.php');
 
 $request = AppContext::get_request();
+$session = AppContext::get_session();
 
 $get_id = $request->get_getint('id', 0);
 $del = $request->get_getint('del', 0);
@@ -27,41 +28,57 @@ $valid = $request->get_postbool('valid', false);
 $view = new FileTemplate('forum/admin_ranks.tpl');
 $view->add_lang($lang);
 
+$csrf_token = $session->get_token();
+
 //Si c'est confirmé on execute
 if ($valid)
 {
-	$result = PersistenceContext::get_querier()->select("SELECT id, special
-	FROM " . PREFIX . "forum_ranks");
-	while ($row = $result->fetch())
-	{
-		$name       = $request->get_poststring($row['id'] . 'name', '');
-		$msg_number = $request->get_postint($row['id'] . 'msg', 0);
-		$icon       = $request->get_poststring($row['id'] . 'icon', '');
+    if ($request->get_poststring('token', '') === $csrf_token)
+    {
+        $result = PersistenceContext::get_querier()->select("SELECT id, special
+        FROM " . PREFIX . "forum_ranks");
+        while ($row = $result->fetch())
+        {
+            $name       = $request->get_poststring($row['id'] . 'name', '');
+            $msg_number = $request->get_postint($row['id'] . 'msg', 0);
+            $icon       = $request->get_poststring($row['id'] . 'icon', '');
 
-		if (!empty($name) && $row['special'] != 1)
-			PersistenceContext::get_querier()->update(PREFIX . "forum_ranks", array('name' => $name, 'msg' => $msg_number, 'icon' => $icon), ' WHERE id = :id', array('id' => $row['id']));
-		else
-			PersistenceContext::get_querier()->update(PREFIX . "forum_ranks", array('name' => $name, 'icon' => $icon), ' WHERE id = :id', array('id' => $row['id']));
-	}
-	$result->dispose();
+            if (!empty($name) && $row['special'] != 1)
+                PersistenceContext::get_querier()->update(PREFIX . "forum_ranks", array('name' => $name, 'msg' => $msg_number, 'icon' => $icon), ' WHERE id = :id', array('id' => $row['id']));
+            else
+                PersistenceContext::get_querier()->update(PREFIX . "forum_ranks", array('name' => $name, 'icon' => $icon), ' WHERE id = :id', array('id' => $row['id']));
+        }
+        $result->dispose();
 
-	ForumRanksCache::invalidate();
+        ForumRanksCache::invalidate();
 
-	HooksService::execute_hook_action('edit_config', 'forum', array('title' => $lang['forum.ranks.management'], 'url' => ForumUrlBuilder::manage_ranks()->rel()));
+        HooksService::execute_hook_action('edit_config', 'forum', array('title' => $lang['forum.ranks.management'], 'url' => ForumUrlBuilder::manage_ranks()->rel()));
 
-	$view->put('MESSAGE_HELPER', MessageHelper::display(LangLoader::get_message('warning.process.success', 'warning-lang'), MessageHelper::SUCCESS, 4));
+        $view->put('MESSAGE_HELPER', MessageHelper::display(LangLoader::get_message('warning.process.success', 'warning-lang'), MessageHelper::SUCCESS, 4));
+    }
+    else
+    {
+        $view->put('MESSAGE_HELPER', MessageHelper::display(LangLoader::get_message('warning.invalid.csrf.token', 'warning-lang'), MessageHelper::ERROR, 4));
+    }
 }
 elseif (!empty($del) && !empty($get_id)) //Suppression du rang.
 {
-	//On supprime dans la bdd.
-	PersistenceContext::get_querier()->delete(PREFIX . 'forum_ranks', 'WHERE id=:id', array('id' => $get_id));
+	if (!empty($token) && hash_equals($csrf_token, $token))
+    {
+        //On supprime dans la bdd.
+        PersistenceContext::get_querier()->delete(PREFIX . 'forum_ranks', 'WHERE id=:id', array('id' => $get_id));
 
-	###### Régénération du cache des rangs #######
-	ForumRanksCache::invalidate();
+        ###### Régénération du cache des rangs #######
+        ForumRanksCache::invalidate();
 
-	HooksService::execute_hook_action('edit_config', 'forum', array('title' => $lang['forum.ranks.management'], 'url' => ForumUrlBuilder::manage_ranks()->rel()));
+        HooksService::execute_hook_action('edit_config', 'forum', array('title' => $lang['forum.ranks.management'], 'url' => ForumUrlBuilder::manage_ranks()->rel()));
 
-	$view->put('MESSAGE_HELPER', MessageHelper::display(LangLoader::get_message('warning.process.success', 'warning-lang'), MessageHelper::SUCCESS, 4));
+        $view->put('MESSAGE_HELPER', MessageHelper::display(LangLoader::get_message('warning.process.success', 'warning-lang'), MessageHelper::SUCCESS, 4));
+    }
+    else
+    {
+        $view->put('MESSAGE_HELPER', MessageHelper::display(LangLoader::get_message('warning.invalid.csrf.token', 'warning-lang'), MessageHelper::ERROR, 4));
+    }
 }
 
 //On recupère les images des groupes
